+++ package/src/dom/__tests__/Demos.spec.tsx
@@ -1,9 +1,6 @@
-import { mapKeys } from "../../renderer/typeddash";
 import { importSkia, width, height } from "../../renderer/__tests__/setup";
-import { setupSkia } from "../../skia/__tests__/setup";
+import { asXML, setupSkia } from "../../skia/__tests__/setup";
 import { processResult } from "../../__tests__/setup";
-import type { Node } from "../types";
-import { isMatrix } from "../../skia/types/Matrix";
 
 describe("Drawings", () => {
   it("Hello World", () => {
@@ -42,48 +39,8 @@ describe("Drawings", () => {
     console.log(asXML(root));
     const ctx = { canvas, paint: Skia.Paint(), opacity: 1, Skia };
     root.render(ctx);
-    processResult(surface, "snapshots/demos/breathe2.png", true);
+    processResult(surface, "snapshots/demos/breathe.png");
   });
 });
 
 // TODO: add test where each circle is not around a group
-
-const asValue = (value: unknown) => {
-  if (value instanceof Float32Array) {
-    return Array.from(value)
-      .map((v) => Math.round(v * 100) / 100)
-      .join(", ");
-  } else if (isMatrix(value)) {
-    // TODO: https://github.com/Shopify/react-native-skia/issues/715
-    return "m3x3";
-  } else if (
-    typeof value === "object" &&
-    value !== null &&
-    "x" in value &&
-    "y" in value
-  ) {
-    const v = value as { x: number; y: number };
-    return `{x:${v.x}, y:${v.y}}`;
-  }
-  return value;
-};
-
-// Print nodes as XML tree
-const asXML = (node: Node<unknown>, indent = 0): string => {
-  // TODO: remove cast
-  const hasChildren = node.children().length > 0;
-  const props = node.getProps() as object;
-  const space = new Array(indent).fill(" ").join("");
-  if (hasChildren) {
-    return `${space}<${node.type} ${mapKeys(props)
-      .map((name) => `${name}="${asValue(props[name])}"`)
-      .join(" ")}>
-${node.children().map((child) => asXML(child, indent + 2)).join(`
-`)}
-${space}</${node.type}>`;
-  } else {
-    return `${space}<${node.type} ${mapKeys(props)
-      .map((name) => `${name}="${asValue(props[name])}"`)
-      .join(" ")} />`;
-  }
-};
