+++ package/src/dom/__tests__/RenderNodes.spec.tsx
@@ -1,0 +1,22 @@
+import { importSkia, width, height } from "../../renderer/__tests__/setup";
+import { setupSkia } from "../../skia/__tests__/setup";
+import { processResult } from "../../__tests__/setup";
+
+describe("Drawings", () => {
+  it("Hello World", () => {
+    const { surface, canvas } = setupSkia(width, height);
+    const { Skia, vec } = importSkia();
+    const c = vec(width / 2, height / 2);
+    const c1 = Skia.Color("#61bea2");
+    //const c2 = Skia.Color("#529ca0");
+    const R = width / 4;
+    //const color = Skia.Color("rgb(36,43,56)");
+    const root = Sk.Group({});
+    root.addChild(Sk.Circle({ c, r: R, color: c1 }));
+    const ctx = { canvas, paint: Skia.Paint(), opacity: 1, Skia };
+    root.render(ctx);
+    processResult(surface, "snapshots/render-nodes/simple.png", true);
+  });
+});
+
+// TODO: add test where each circle is not around a group
