+++ package/src/skia/__tests__/setup.ts
@@ -1,5 +1,7 @@
+import type { Node } from "../../dom/types";
+import { mapKeys } from "../../renderer/typeddash";
 import { LoadSkiaWeb } from "../../web/LoadSkiaWeb";
-import { Skia } from "../types";
+import { isMatrix, Skia } from "../types";
 import { JsiSkApi } from "../web";
 
 let Skia: ReturnType<typeof JsiSkApi>;
@@ -24,4 +26,44 @@ export const setupSkia = (width = 256, height = 256) => {
     Skia,
     CanvasKit: global.CanvasKit,
   };
+};
+
+const asValue = (value: unknown) => {
+  if (value instanceof Float32Array) {
+    return Array.from(value)
+      .map((v) => Math.round(v * 100) / 100)
+      .join(", ");
+  } else if (isMatrix(value)) {
+    // TODO: https://github.com/Shopify/react-native-skia/issues/715
+    return "m3x3";
+  } else if (
+    typeof value === "object" &&
+    value !== null &&
+    "x" in value &&
+    "y" in value
+  ) {
+    const v = value as { x: number; y: number };
+    return `{x:${v.x}, y:${v.y}}`;
+  }
+  return value;
+};
+
+// Print nodes as XML tree
+export const asXML = (node: Node<unknown>, indent = 0): string => {
+  // TODO: remove cast
+  const hasChildren = node.children().length > 0;
+  const props = node.getProps() as object;
+  const space = new Array(indent).fill(" ").join("");
+  if (hasChildren) {
+    return `${space}<${node.type} ${mapKeys(props)
+      .map((name) => `${name}="${asValue(props[name])}"`)
+      .join(" ")}>
+${node.children().map((child) => asXML(child, indent + 2)).join(`
+`)}
+${space}</${node.type}>`;
+  } else {
+    return `${space}<${node.type} ${mapKeys(props)
+      .map((name) => `${name}="${asValue(props[name])}"`)
+      .join(" ")} />`;
+  }
 };
